Dockerfile, env mapping, IP overrides
This commit adds an initial Dockerfile, however
more tweaking is needed for it.

Environment overrides are also added for a number
of configuration settings. This is primarily
useful with Docker.

Added setting for IP overriding for the harvester
HTTP server. This is mainly needed when using
Docker and having traffic come from outside. There
could be ways to self detect, but it doesn't work
with Docker on Mac. That likely isn't the main
expected use case though.

Added fixed number of transfer routines for the
plotter command, commandline configurable with a
default of 2.

Added gomaxproc instrumentation.