harvester: Added graceful shutdown
This adds a graceful shutdown on the harvester
handling. It will immediately close the NATS
connection and shutdown the http server listening
to new connections, however it will allow existing
transfers to complete. Note that you may want to
ensure a high enough shutdown timeout in Docker or
whereever else relative to your plot file size and
network transfer capacity.