harvester: Add expand-path option
This adds the --expand-path option for the harvester which will look at
the provided path and use each subdirectory within the directory as a
separate plot directory. This is a simple way to have one argument
point to a directory containing the mount point of all disks and
discovering all of them instantly.