harvester: Pause plot paths on transfer failures
The harvester will now pause a specific plot path for 5 minutes if the
write or open operations fails. These are the two cases where the
failure is most likely related to the disk/filesystem and could warrant
skipping it so as to not get stuck.

Future changes should involve periodically updating stats and disabling
on a failure.