harvester: Use temp file and rename when complete
This changes the harvester to write the files to a temp file while
transfering and rename them upon completion. Should have been doing this
already, but neglected it. The main benefit is the chia harvester won't
see or try using it until the transfer is complete. Renaming is a quick
syscall so it is fully ready once the farm sees it.