fix: Miscalculating of free space taint
This fixes two issues with calculating the free space taint.

First, it was calculating it inverted. It was using free space
percentage rather than disk usage percentage, so it was tainting more
the emptier the disk was.

Second, it was treating the taint as 1/10th what it should because it
was multiplying by 100 and then dividing by 1000, so it was off by a
factor of 10.